+++ cmp/dataview/DataViewModel.js
@@ -33,10 +33,9 @@ export class DataViewModel {
      * @param {(StoreSelectionModel|Object|String)} [c.selModel] - StoreSelectionModel, or a
      *      config or string `mode` from which to create.
      * @param {string} [c.emptyText] - text/HTML to display if view has no records.
-     * @param {function} [c.contextMenuFn] - closure returning a StoreContextMenu().
-     * @param {(Object[]|function)} [c.contextMenu] - array of RecordActions, configs or token strings with which to create
-     *      grid context menu items or function to optionally return a
-     *      StoreContextMenu when the grid is right-clicked (desktop only).
+     * @param {(Object[]|GridStoreContextMenuFn)} [c.contextMenu] - array of RecordActions, configs or token
+     *      strings with which to create grid context menu items.  May also be specified as a
+     *      function returning a StoreContextMenu.  Desktop only.
      */
     constructor({
         itemRenderer,
@@ -44,8 +43,7 @@ export class DataViewModel {
         sortBy = [],
         selModel,
         emptyText,
-        contextMenu = null,
-        contextMenuFn = null
+        contextMenu = null
     }) {
         sortBy = castArray(sortBy);
         throwIf(sortBy.length > 1, 'DataViewModel does not support multiple sorters.');
@@ -67,7 +65,6 @@ export class DataViewModel {
             sortBy,
             selModel,
             contextMenu,
-            contextMenuFn,
             emptyText,
             columns: [
                 {